✨ 関数をファイル分けした&makefileとminishell.hも修正
