+++ Makefile
@@ -6,7 +6,7 @@
 #    By: miyuu <miyuu@student.42.fr>                +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/05/09 00:35:59 by tkondo            #+#    #+#              #
-#    Updated: 2025/03/03 19:37:45 by miyuu            ###   ########.fr        #
+#    Updated: 2025/03/03 20:00:49 by miyuu            ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -51,6 +51,10 @@ TARGET =\
 	data/load_simple_cmd\
 	data/pipe2scmd_list\
 	data/tokenizer_scmd_text\
+	data/add_struct_text_list\
+	data/ft_isspace\
+	data/get_redir_length\
+	data/new_struct_text_list\
 	expand/expand_ecmds\
 	expand/get_exit_status\
 	expand/get_exit_status_p\
