+++ src/data/add_struct_text_list.c
@@ -1,0 +1,33 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   add_struct_text_list.c                             :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: miyuu <miyuu@student.42.fr>                +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/03/03 19:56:26 by miyuu             #+#    #+#             */
+/*   Updated: 2025/03/03 19:56:34 by miyuu            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include <minishell.h>
+
+/*
+ * Function:add_struct_text_list
+ * ----------------------------
+ */
+void	add_struct_text_list(t_text_list **head, t_text_list *new)
+{
+	t_text_list	*cur;
+
+	if (!*head)
+		*head = new;
+	else
+	{
+		cur = *head;
+		while (cur->next)
+			cur = cur->next;
+		cur->next = new;
+		new->prev = cur;
+	}
+}
