+++ src/data/fill_ecmds.c
@@ -18,50 +18,12 @@
  * Returns an array only strings without redirects.
  * ToDO:norminetteエラー
  */
-// char	**fill_ecmds(char **src, int wc)
-// {
-// 	char	**ecmds;
-// 	int		i;
-// 	int		j;
-// 	int		len;
-
-// 	i = 0;
-// 	j = 0;
-// 	ecmds = (char **)malloc(sizeof(char *) * (wc + 1));
-// 	if (!ecmds)
-// 		return (NULL);
-// 	while (src[i])
-// 	{
-// 		if (has_redirect(src[i]) != NULL)
-// 		{
-// 			//ToDo:リダイレクトを含む文字列の最後もの字が記号かいなか関数分けする？
-// 			len = ft_strlen(src[i]);
-// 			if (src[i + 1] != NULL && \
-// 				(src[i][len - 1] == '>' || src[i][len - 1] == '<'))
-// 				i++;
-// 		}
-// 		else
-// 		{
-// 			ecmds[j] = ft_strdup(src[i]);
-// 			if (!ecmds[j])
-// 			{
-// 				free_ecmds(ecmds);
-// 				return (NULL);
-// 			}
-// 			j++;
-// 		}
-// 		i++;
-// 	}
-// 	ecmds[j] = NULL;
-// 	return (ecmds);
-// }
-
 char	**fill_ecmds(t_text_list *scmds, int wc)
 {
-	char	**ecmds;
-	int		j;
-	int		len;
-	t_text_list			*cur;
+	char		**ecmds;
+	int			j;
+	int			len;
+	t_text_list	*cur;
 
 	j = 0;
 	ecmds = (char **)malloc(sizeof(char *) * (wc + 1));
